New base patch
Gives all dungeons unique names in the crossroads menu.
A star indicates the dungeon is from a Special Episode. An up/down arrow means there are two segments of the dungeon with the exact same name in a row, with the up arrow representing the first segment and the down arrow representing the second segment. Also fixed issues causing softlocks after dungeons by effectively starting the player in postgame. And an easter egg :P